update(lambda): meta schema parsing
